1 - Q&A Chatbot/requirements.txt
langchain_openai
langchain
python-dotenv
langchain_core
langchain_community
streamlit